fragAlfonz/backupImageGen/AlfonzImageGen/requirements.txt
torch==2.5.1
diffusers==0.30.3
transformers==4.46.3
accelerate==1.2.1
pillow==11.1.0
safetensors==0.5.3
tokenizers==0.20.3
fastapi
uvicorn